+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.0.2 - 11/01/2022
+### Breaking
+* Changed params of the file store manager to `key` instead of filename
+* removed the deleteAll from the file store manager
+
 ## 0.0.1 - 11/01/2022
 * initial release